Module finchers_juniper::execute

GraphQL executors.

In order to choose the strategy for executing GraphQL queries, the following GraphQL executors are provided:

Nonblocking It spawns the task for executing GraphQL queries by using the Tokio's default executor. It also notify the start of the blocking section to the runtime by using the blocking API provided by tokio_threadpool.

CurrentThread It executes the GraphQL queries on the current thread.

WithSpawner It spawn the task for executing GraphQL queries by using the specified task executor. Unlike to Nonblocking, it does not notify the start of blocking section to the runtime.

Structs

CurrentThread

Nonblocking

WithSpawner

Traits

Schema
Trait representing a GraphQL schema.

SharedSchema
A helper trait for representing a Schema which can be shared between threads.

Functions

current_thread
Create a GraphQL executor from the specified RootNode.

nonblocking
Create a GraphQL executor from the specified RootNode.

with_spawner
Create a GraphQL executor from the specified RootNode and task executor.