Function finchers_juniper::execute::with_spawner

pub fn with_spawner<S, Sp>(schema: S, spawner: Sp) -> WithSpawner<S, Sp> where
    S: SharedSchema,
    Sp: Executor<Execute<GraphQLTask<S>>>, 

Create a GraphQL executor from the specified RootNode and task executor.

The endpoint created by this wrapper will spawn a task which executes the GraphQL queries after receiving the request, by using the specified Executor<T>.